aqia/server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
coqui-tts
python-multipart==0.0.6
groq==0.4.2
transformers==4.33.0
google-cloud-texttospeech
python-dotenv
pytest
httpx
